requirements.txt
django-environ==0.12.0
sslcommerz-python-api==1.0.1
